remove old onboarding / welcome
