--- assets/css/onboard.css
@@ -1,23 +1,0 @@
-#prpl-onboarding-form .prpl-form-fields label {
-	display: grid;
-	grid-template-columns: 1fr 3fr;
-	margin-bottom: 0.5em;
-	gap: var(--prpl-padding);
-}
-
-#prpl-onboarding-form label > span:has(input[type="checkbox"]) {
-	display: flex;
-	align-items: baseline;
-}
-
-.prpl-onboard-form-radio-select {
-
-	label {
-		display: block !important;
-	}
-}
-
-#prpl-onboarding-submit-wrapper {
-	display: flex;
-	align-items: center;
-}
